new version
new version with still the bug, that not all is committed correctly

+++ CostFunction/CostFunction/AngleGenerator.h
@@ -3,7 +3,7 @@ class AngleGenerator
 {
 public:
 	AngleGenerator();
-	AngleGenerator(float* costs, int nAngle, int DOF);
+	AngleGenerator(double* costs, int nAngle, int DOF);
 	~AngleGenerator(void);
 
 	int generateRandomAngle();
